+++ apps/setup/steps.c
@@ -1410,7 +1410,7 @@ Return Value:
             if (BytesDone != SETUP_BLOCK_SIZE) {
                 fprintf(stderr,
                         "Read only %ld of %d bytes.\n",
-                        BytesDone,
+                        (long)BytesDone,
                         SETUP_BLOCK_SIZE);
 
                 Result = errno;
@@ -1530,7 +1530,7 @@ Return Value:
         if (BytesDone != SETUP_BLOCK_SIZE) {
             fprintf(stderr,
                     "Error: Wrote only %ld of %d bytes.\n",
-                    BytesDone,
+                    (long)BytesDone,
                     SETUP_BLOCK_SIZE);
 
             Result = errno;
