+++ lib/rtl/testrtl/testrtl.c
@@ -711,7 +711,7 @@ Return Value:
         printf("Error: Print basic string with NULL output returned output "
                "length of %d, should have been %lu.\n",
                StringLength,
-               strlen(BASIC_STRING) + 1);
+               (long)strlen(BASIC_STRING) + 1);
 
         TestsFailed += 1;
     }
@@ -729,7 +729,7 @@ Return Value:
         printf("Error: Print basic string with no output returned output "
                "length of %d, should have been %lu.\n",
                StringLength,
-               strlen(BASIC_STRING) + 1);
+               (long)strlen(BASIC_STRING) + 1);
 
         TestsFailed += 1;
     }
@@ -755,7 +755,7 @@ Return Value:
         printf("Error: Print formatted string with no output returned output "
                "length of %d, should have been %lu.\n",
                StringLength,
-               strlen(FORMATTED_STRING_RESULT) + 1);
+               (long)strlen(FORMATTED_STRING_RESULT) + 1);
 
         TestsFailed += 1;
     }
@@ -769,7 +769,7 @@ Return Value:
         printf("Error: Print formatted string with output returned output "
                "length of %d, should have been %lu.\n",
                StringLength,
-               strlen(FORMATTED_STRING_RESULT) + 1);
+               (long)strlen(FORMATTED_STRING_RESULT) + 1);
 
         TestsFailed += 1;
     }
@@ -814,7 +814,7 @@ Return Value:
         printf("Error: Print formatted string with output returned output "
                "length of %d, should have been %lu.\n",
                StringLength,
-               strlen(FORMATTED_STRING_POSITIONAL_RESULT) + 1);
+               (long)strlen(FORMATTED_STRING_POSITIONAL_RESULT) + 1);
 
         TestsFailed += 1;
     }
@@ -863,7 +863,7 @@ Return Value:
         printf("Error: Print float string with output returned output "
                "length of %d, should have been %lu.\n",
                StringLength,
-               strlen(PRINT_FLOAT_RESULT) + 1);
+               (long)strlen(PRINT_FLOAT_RESULT) + 1);
 
         TestsFailed += 1;
     }
@@ -891,7 +891,7 @@ Return Value:
         printf("Error: Print hex float string with output returned output "
                "length of %d, should have been %lu.\n",
                StringLength,
-               strlen(PRINT_HEX_FLOAT_RESULT) + 1);
+               (long)strlen(PRINT_HEX_FLOAT_RESULT) + 1);
 
         TestsFailed += 1;
     }
