+++ uefi/tools/mkuboot/mkuboot.c
@@ -618,7 +618,7 @@ Return Value:
                 "mkuboot: Unable to read \"%s\". Read %ld bytes, expected "
                 "%d.\n",
                 Context->InputFileName,
-                BytesRead,
+                (long)BytesRead,
                 InputFileSize);
 
         Result = 1;
@@ -778,10 +778,10 @@ Return Value:
 
     if (BytesWritten != sizeof(UBOOT_HEADER)) {
         fprintf(stderr,
-                "mkuboot: Needed to write %ld byte U-Boot header, wrote %ld "
+                "mkuboot: Needed to write %d byte U-Boot header, wrote %ld "
                 "bytes.\n",
-                sizeof(UBOOT_HEADER),
-                BytesWritten);
+                (int)sizeof(UBOOT_HEADER),
+                (long)BytesWritten);
 
         Result = 1;
         goto CreateLegacyImageEnd;
@@ -928,7 +928,7 @@ Return Value:
         fprintf(stderr,
                 "mkuboot: Failed to write FIT strings. Write %ld bytes, "
                 "expected %d.\n",
-                BytesWritten,
+                (long)BytesWritten,
                 StringsSize);
 
         Result = 1;
@@ -990,10 +990,10 @@ Return Value:
 
     if (BytesWritten != sizeof(UBOOT_FIT_HEADER)) {
         fprintf(stderr,
-                "mkuboot: Needed to write %ld byte U-Boot FIT header, "
+                "mkuboot: Needed to write %d byte U-Boot FIT header, "
                 "wrote %ld bytes.\n",
-                sizeof(UBOOT_FIT_HEADER),
-                BytesWritten);
+                (int)sizeof(UBOOT_FIT_HEADER),
+                (long)BytesWritten);
 
         Result = 1;
         goto CreateFitImageEnd;
